Refactor FileHighlighter.java
